General data control improvements
Resolved several deep issues in RSObject data calculation. This seems to hve resulted in a single update cycle however this currently neglects deep referencing in calculated fields, thoough a use case to support this seems weak at the moment.

In either case, the calculation logic for RSObject is in need of cleaning but currently seems functional. Will continue usage and testing and wait till the process feels stable and better known.