Authentication fix, date handling, calculator
Additional minor fixes, copying adjusted to use new "is_singular" field to prevent copying and resolved an authentication bug that allowed preview classed players to login.

Started a new class in the definitions for previews as a general case to get them out of the primary class buckets as well.